# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Die Groot Ompad — Integrated Travel Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List all trips
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Trip created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{tripId}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get a trip by ID
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /trips/{tripId}/summary:
    get:
      operationId: getTripSummary
      tags: [analytics]
      summary: Get aggregated summary metrics for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripSummary"

  # ── Legs ──────────────────────────────────────────────────────────────────
...
  /trips/{tripId}/vehicle:
    get:
      operationId: getVehicleProfile
      tags: [vehicle]
      summary: Get vehicle and caravan profile for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleProfile"
    put:
      operationId: saveVehicleProfile
      tags: [vehicle]
      summary: Save vehicle and caravan profile
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleProfileInput"
      responses:
        "200":
          description: Saved vehicle profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleProfile"

  # ── GPS ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Trip:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        notes:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        fuelPrice15:
          type: number
        fuelPrice18:
          type: number
        fuelPrice20:
          type: number
        revision:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    TripInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        notes:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        fuelPrice15:
          type: number
        fuelPrice18:
          type: number
        fuelPrice20:
          type: number
    TripUpdate:
      type: object
      properties:
        name:
          type: string
        notes:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        fuelPrice15:
          type: number
        fuelPrice18:
          type: number
        fuelPrice20:
          type: number
        revision:
          type: string
    TripSummary:
      type: object
      required: [tripId, totalLegs, totalPlannedKm, totalActualKm, totalEstFuelCost18, totalActualFuelCost, avgConsumptionL100km, kmVariance, costVariance]
      properties:
        tripId:
          type: integer
        totalLegs:
          type: integer
        totalPlannedKm:
          type: number
        totalActualKm:
          type: number
        totalEstFuelCost18:
          type: number
        totalActualFuelCost:
          type: number
        avgConsumptionL100km:
          type: number
        kmVariance:
          type: number
        costVariance:
          type: number

    # ── Leg ───────────────────────────────────────────────────────────────────
...
    VehicleProfile:
      type: object
      required: [id, tripId, updatedAt]
      properties:
        id:
          type: integer
        tripId:
          type: integer
        vehicleModel:
          type: string
        vehicleFuel:
          type: string
        kerbWeight:
          type: number
        gvm:
          type: number
        gcm:
          type: number
        towRating:
          type: number
        frontAxleLimit:
          type: number
        rearAxleLimit:
          type: number
        caravanModel:
          type: string
        caravanType:
          type: string
        caravanTare:
          type: number
        caravanAtm:
          type: number
        caravanGtm:
          type: number
        ballWeight:
          type: number
        waterLoad:
          type: number
        extrasLoad:
          type: number
        payloadPeople:
          type: number
        payloadFood:
          type: number
        payloadRecovery:
          type: number
        payloadTools:
          type: number
        payloadFuel:
          type: number
        payloadOther:
          type: number
        updatedAt:
          type: string
    VehicleProfileInput:
      type: object
      properties:
        vehicleModel:
          type: string
        vehicleFuel:
          type: string
        kerbWeight:
          type: number
        gvm:
          type: number
        gcm:
          type: number
        towRating:
          type: number
        frontAxleLimit:
          type: number
        rearAxleLimit:
          type: number
        caravanModel:
          type: string
        caravanType:
          type: string
        caravanTare:
          type: number
        caravanAtm:
          type: number
        caravanGtm:
          type: number
        ballWeight:
          type: number
        waterLoad:
          type: number
        extrasLoad:
          type: number
        payloadPeople:
          type: number
        payloadFood:
          type: number
        payloadRecovery:
          type: number
        payloadTools:
          type: number
        payloadFuel:
          type: number
        payloadOther:
          type: number

    # ── GPS ────────────────────────────────────────────────────────────────────
...