# lib/api-zod/src/generated/types/trip.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Die Groot Ompad — Integrated Travel Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Trip {
  id: number;
  name: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  fuelPrice15?: number;
  fuelPrice18?: number;
  fuelPrice20?: number;
  revision?: string;
  createdAt: string;
  updatedAt: string;
}