# lib/api-zod/src/generated/types/tripUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Die Groot Ompad — Integrated Travel Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TripUpdate {
  name?: string;
  notes?: string;
  startDate?: string;
  endDate?: string;
  fuelPrice15?: number;
  fuelPrice18?: number;
  fuelPrice20?: number;
  revision?: string;
}