# lib/api-zod/src/generated/types/vehicleProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Die Groot Ompad — Integrated Travel Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface VehicleProfile {
  id: number;
  tripId: number;
  vehicleModel?: string;
  vehicleFuel?: string;
  kerbWeight?: number;
  gvm?: number;
  gcm?: number;
  towRating?: number;
  frontAxleLimit?: number;
  rearAxleLimit?: number;
  caravanModel?: string;
  caravanType?: string;
  caravanTare?: number;
  caravanAtm?: number;
  caravanGtm?: number;
  ballWeight?: number;
  waterLoad?: number;
  extrasLoad?: number;
  payloadPeople?: number;
  payloadFood?: number;
  payloadRecovery?: number;
  payloadTools?: number;
  payloadFuel?: number;
  payloadOther?: number;
  updatedAt: string;
}